Refactor traffic-friendly handling of invisible layers
Before, the whole layer container was removed from the stack's view when
a layer was made invisible and it was re-added when it was made visible
again. While this works in principle, it allowed layers to be added in
the wrong order. Additionally, removing and adding elements causes a
reflow of the page. Now, only the src attributes of the images are
emptied when a layer is made invisible. In that case new tiles don't get
assigned a src attribute in the first place. The actual URL is always
kept as data property of the image object and can thereby easily
retrieved. This is related to issue #569.