Open new stacks of same project with last zoom level
So far a stack's default zoom level was used if it was opened along an
already open stack in the front-end. This commit changes this behavior
to use the zoom level of the stack focussed last for the new stack, if it is
available. This still isn't the best solution, because a new stack might have a
different resolution and the notion of the scale doesn't really reflect this.

However, in the multi-view-tracing branch this has been dealt with and as soon
as it will be merged a resolution is used for this instead of a scale level.

This is related to #790.